% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_titan_binary}
\alias{plot_titan_binary}
\title{Plot binary PLS-LDA outputs}
\usage{
plot_titan_binary(predictions)
}
\arguments{
\item{predictions}{Output from [predict_titan()] for one patient.}
}
\value{
A ggplot object. Scores are displayed as TCGA out-of-fold score
  ranks explicitly labelled "not probability".
}
\description{
Plot binary PLS-LDA outputs
}
