% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_store.R
\name{validate_pathofmpred_object}
\alias{validate_pathofmpred_object}
\title{Validate a PathoFMPred model collection}
\usage{
validate_pathofmpred_object(object)
}
\arguments{
\item{object}{An object created by [create_pathofmpred_object()] or a
distributed PathoFMPred model collection.}
}
\value{
`TRUE`, invisibly, when the object is valid.
}
\description{
Validate a PathoFMPred model collection
}
